Fixed: Win32 DX video rendering is not access from DVDPlayer anymore, eliminating random crashes The WinRenderer is initialized correctly, not crushing when reopening a movie.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@22951 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ guilib/GraphicContext.cpp
@@ -227,7 +227,7 @@ bool CGraphicContext::SetViewPort(float fx, float fy, float fwidth, float fheigh
   ASSERT(newLeft < newRight);
   ASSERT(newTop < newBottom);
 
-  CRect newviewport(newLeft, newTop, newRight, newBottom);
+  CRect newviewport((float)newLeft, (float)newTop, (float)newRight, (float)newBottom);
   g_Windowing.SetViewPort(newviewport);
 
   m_viewStack.push(oldviewport);
